% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/institutions.R
\name{make_question_table}
\alias{make_question_table}
\title{Make a table with the theme and the title of each question}
\usage{
make_question_table(x)
}
\arguments{
\item{x}{a list at the theme level}
}
\value{
a tibble
}
\description{
Make a table with the theme and the title of each question
}
\examples{
library(xml2)
library(dplyr)
library(purrr)
list_institutions <- read_xml(x = "raw-data/Institutions.xml") \%>\% as_list()
make_question_table(list_institutions$themes[[1]]) \%>\% glimpse()
map_df(list_institutions$themes, make_question_table) \%>\% glimpse()
}

