% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/institutions.R
\name{make_questions_reponses_table}
\alias{make_questions_reponses_table}
\title{Make a table with all answers to all questions in a theme}
\usage{
make_questions_reponses_table(x)
}
\arguments{
\item{x}{}
}
\value{
a tibble
}
\description{
Make a table with all answers to all questions in a theme
}
\examples{
library(xml2)
library(dplyr)
library(purrr)
list_institutions <- read_xml(x = "raw-data/Institutions.xml") \%>\% as_list()
make_questions_reponses_table(x = list_institutions$themes[[1]]) \%>\% glimpse()
map_df(.x = list_institutions$themes, make_questions_reponses_table) \%>\% glimpse()
}

