% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/institutions.R
\name{make_reponse_table}
\alias{make_reponse_table}
\title{Make a table from a response list}
\usage{
make_reponse_table(x)
}
\arguments{
\item{x}{}
}
\value{
a tibble with two columns : reponse and n
}
\description{
Make a table from a response list
}
\examples{
library(xml2)
library(dplyr)
library(purrr)
list_institutions <- read_xml(x = "raw-data/Institutions.xml") \%>\% as_list()
make_reponse_table(list_institutions$themes[[1]]$questions[[1]]$reponses[[1]])
}

