% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egalite_homme_femme.R
\name{make_table_egalite_question}
\alias{make_table_egalite_question}
\title{Create a table with all answers to a question}
\usage{
make_table_egalite_question(list)
}
\arguments{
\item{list}{}
}
\value{
a tibble
}
\description{
Create a table with all answers to a question
}
\examples{
make_table_egalite_question(extract2(extract2(json_egalite, "questions"), 1))  \%>\% glimpse()
table_questions_reponses <- map_df(.x = extract2(json_egalite, "questions"), make_table_egalite_question, .id = "id_enquete")
}

