% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egalite_homme_femme.R
\name{make_table_egalite_reponses}
\alias{make_table_egalite_reponses}
\title{Transform answers to a question into a tibble}
\usage{
make_table_egalite_reponses(list)
}
\arguments{
\item{list}{}
}
\value{
a tibble
}
\description{
Transform answers to a question into a tibble
}
\examples{
make_table_egalite_reponses(extract2(extract2(json_egalite, "questions"), 1))
}

