% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/europe.R
\name{make_table_europe}
\alias{make_table_europe}
\title{Make the table woth the first 5 questions and corresponding answers}
\usage{
make_table_europe()
}
\value{
a tibble
}
\description{
Make the table woth the first 5 questions and corresponding answers
}
\examples{
table_europe <- make_table_europe()
}

