% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/europe.R
\name{make_table_question_reponse_europe}
\alias{make_table_question_reponse_europe}
\title{Make a table with the question and all answers}
\usage{
make_table_question_reponse_europe(list, i)
}
\arguments{
\item{list}{the name of the list}

\item{i}{an index}
}
\value{
a tibble
}
\description{
Make a table with the question and all answers
}
\examples{
make_table_question_reponse_europe(i = 1)
}

